ai_bail_agent/requirements.txt
langchain
langchain-core
langchain-openai
langchain-anthropic
langgraph
python-dotenv
ipython
pydantic
langsmith
